# openapi.yaml
components:
  headers:
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        minimum: 1
        type: integer
  parameters:
...
    GameId:
      description: The game, as carried on gameStart and moveRequest.
      in: path
      name: gameId
      required: true
      schema:
        type: string
  responses:
...
    Delisted:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DelistedError"
      description: The bot is delisted and takes no new games.
    GameOver:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GameOverError"
      description: The game is already finished (game_over).
    NotFound:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFoundError"
      description: The target does not exist or is not the caller's to act on.
...
    Paused:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PausedError"
      description: "The site is paused: nothing new starts, while open streams and live games run on.
        Retry after 60 s."
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UnauthorizedError"
      description: No session cookie, or the session is expired or unknown.
  schemas:
...
    DelistedError:
      properties:
        code:
          enum:
            - delisted
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    FinishReason:
      enum:
        - aborted
        - disconnect
        - surrender
        - timeout
        - terminated
        - six-in-a-row
      type: string
    GameBoard:
      properties:
        cells:
          items:
            properties:
              side:
                enum:
                  - x
                  - o
                type: string
              x:
                type: integer
              y:
                type: integer
            required:
              - x
              - y
              - side
            type: object
          type: array
      required:
        - cells
      type: object
    GameClock:
      oneOf:
        - properties:
            mode:
              enum:
                - unlimited
              type: string
          required:
            - mode
          type: object
        - properties:
            mode:
              enum:
                - turn
              type: string
            remainingTurnMs:
              minimum: 0
              type: integer
          required:
            - mode
            - remainingTurnMs
          type: object
        - properties:
            mode:
              enum:
                - match
              type: string
            remainingMainMs:
              properties:
                o:
                  minimum: 0
                  type: integer
                x:
                  minimum: 0
                  type: integer
              required:
                - x
                - o
              type: object
          required:
            - mode
            - remainingMainMs
          type: object
    GameCreateError:
      properties:
        code:
          enum:
            - bad_request
            - human_busy
            - game_cooldown
            - not_open
            - clock_not_accepted
            - bot_busy
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
    GameEvent:
      description: "One server-sent event: the event field names the kind, and the data field holds the
        payload as one line of JSON."
      oneOf:
        - properties:
            data:
              $ref: "#/components/schemas/GameSnapshot"
            event:
              enum:
                - snapshot
              type: string
          required:
            - event
            - data
          type: object
        - properties:
            data:
              $ref: "#/components/schemas/GameTurn"
            event:
              enum:
                - turn
              type: string
          required:
            - event
            - data
          type: object
        - properties:
            data:
              $ref: "#/components/schemas/GameFinish"
            event:
              enum:
                - finish
              type: string
          required:
            - event
            - data
          type: object
    GameFinish:
      description: The result, with the clock as it stood at the end.
      properties:
        clock:
          $ref: "#/components/schemas/GameClock"
        reason:
          $ref: "#/components/schemas/FinishReason"
        winner:
          enum:
            - x
            - o
            - null
          nullable: true
          type: string
      required:
        - winner
        - reason
        - clock
      type: object
...
    GameOverError:
      properties:
        code:
          enum:
            - game_over
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
    GamePlayer:
      properties:
        kind:
          enum:
            - bot
            - user
            - guest
          type: string
        name:
          type: string
        provisional:
          $ref: "#/components/schemas/Provisional"
        rating:
          allOf:
            - $ref: "#/components/schemas/Rating"
            - nullable: true
      required:
        - name
        - rating
        - provisional
        - kind
      type: object
    GamePlayers:
      properties:
        o:
          $ref: "#/components/schemas/GamePlayer"
        x:
          $ref: "#/components/schemas/GamePlayer"
      required:
        - x
        - o
      type: object
    GameSnapshot:
      description: you is the caller's side, present exactly when the caller holds a seat. board.cells
        lists stones in ply order, so its first openingPlies entries are the opening.
      oneOf:
        - properties:
            board:
              $ref: "#/components/schemas/GameBoard"
            clock:
              $ref: "#/components/schemas/GameClock"
            gameId:
              type: string
            openingPlies:
              $ref: "#/components/schemas/OpeningPlies"
            players:
              $ref: "#/components/schemas/GamePlayers"
            status:
              enum:
                - in-progress
              type: string
            toMove:
              enum:
                - x
                - o
              type: string
            you:
              enum:
                - x
                - o
              type: string
          required:
            - gameId
            - players
            - openingPlies
            - board
            - status
            - toMove
            - clock
          type: object
        - properties:
            board:
              $ref: "#/components/schemas/GameBoard"
            clock:
              $ref: "#/components/schemas/GameClock"
            gameId:
              type: string
            openingPlies:
              $ref: "#/components/schemas/OpeningPlies"
            players:
              $ref: "#/components/schemas/GamePlayers"
            reason:
              $ref: "#/components/schemas/FinishReason"
            status:
              enum:
                - finished
              type: string
            winner:
              enum:
                - x
                - o
                - null
              nullable: true
              type: string
            you:
              enum:
                - x
                - o
              type: string
          required:
            - gameId
            - players
            - openingPlies
            - board
            - status
            - winner
            - reason
          type: object
...
    GameTurn:
      description: One applied turn, numbered from turn 0 at the origin. cells holds its stones in order;
        a turn whose first stone wins holds that stone alone. toMove and clock are the game's state
        once the turn lands.
      properties:
        cells:
          items:
            properties:
              x:
                type: integer
              y:
                type: integer
            required:
              - x
              - y
            type: object
          maxItems: 2
          minItems: 1
          type: array
        clock:
          $ref: "#/components/schemas/GameClock"
        side:
          enum:
            - x
            - o
          type: string
        toMove:
          enum:
            - x
            - o
          type: string
        turn:
          minimum: 1
          type: integer
      required:
        - turn
        - side
        - cells
        - toMove
        - clock
      type: object
...
    LiveGameEntry:
      description: A game in progress; a game with a guest seat is unrated.
      properties:
        gameId:
          type: string
        players:
          $ref: "#/components/schemas/GamePlayers"
        plies:
          description: Stones on the board, the opening included.
          minimum: 1
          type: integer
        rated:
          type: boolean
        timeControl:
          $ref: "#/components/schemas/TimeControl"
        toMove:
          enum:
            - x
            - o
          type: string
      required:
        - gameId
        - players
        - timeControl
        - toMove
        - rated
        - plies
      type: object
    MoveError:
      properties:
        code:
          enum:
            - bad_request
            - not_your_turn
            - cell_occupied
            - out_of_range
            - game_over
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    NotFoundError:
      properties:
        code:
          enum:
            - not_found
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    Ok:
      properties:
        ok:
          enum:
            - true
          type: boolean
      required:
        - ok
      type: object
    OpeningPlies:
      description: Stones the server places before either player acts, the origin at (0, 0) included. A
        ply is one stone; turn 0 is the origin, and turn t >= 1 is plies 2t-1 and 2t. Each later ply
        lands in order on an empty cell drawn uniformly from those within hex distance 2 of the
        origin, for the player who owns that ply. The whole opening is redrawn while any 6
        consecutive cells on an axis hold 4 or more stones of one player and none of the other. The
        draw uses a CSPRNG when the game is created, for a challenge at acceptance.
      enum:
        - 1
        - 3
        - 5
        - 7
        - 9
      type: integer
    PausedError:
      properties:
        code:
          enum:
            - paused
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    Provisional:
      description: True while the rating deviation is above 75.
      type: boolean
    Rating:
      description: The Glicko-2 rating, rounded to a whole point.
      type: integer
...
    TimeControl:
      description: Durations are in milliseconds.
      oneOf:
        - properties:
            mode:
              enum:
                - unlimited
              type: string
          required:
            - mode
          type: object
        - properties:
            mode:
              enum:
                - turn
              type: string
            turnTimeMs:
              minimum: 5000
              type: integer
          required:
            - mode
            - turnTimeMs
          type: object
        - properties:
            incrementMs:
              minimum: 0
              type: integer
            mainTimeMs:
              minimum: 60000
              type: integer
            mode:
              enum:
                - match
              type: string
          required:
            - mode
            - mainTimeMs
            - incrementMs
          type: object
    UnauthorizedError:
      properties:
        code:
          enum:
            - unauthorized
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
...
    WatcherLimitError:
      properties:
        code:
          enum:
            - watcher_limit
          type: string
        error:
          type: string
      required:
        - error
        - code
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: opaque
      description: A bot token (hxo_...), minted and rotated by the owner on the website.
      scheme: bearer
      type: http
    gameToken:
      bearerFormat: opaque
      description: A game token (hgs_...) from a gameStart line.
      scheme: bearer
      type: http
    sessionCookie:
      description: An HttpOnly session cookie, set by the Discord OAuth callback or the guest route.
      in: cookie
      name: hexo_arena_session
      type: apiKey
info:
  title: HeXO Arena
  version: 0.19.0
openapi: 3.0.3
paths:
...
  /api/bot/game/{gameId}/resign:
    post:
      description: The opponent wins with reason surrender; no engine session is needed.
      operationId: resignBotGame
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
          description: Resigned.
        "400":
          $ref: "#/components/responses/GameOver"
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Missing, unknown, expired, or rotated game token.
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - gameToken: []
      summary: Resign a game over the engine-session token.
      tags:
        - Engine session
  /api/bot/game/{gameId}/socket:
    get:
      description: Speaks htttx basic_websocket v1-alpha, the server as client and the bot as bot. The bot
        must support the move_skips and request_id capabilities. A new connection replaces the
        previous one. Each move_request lists in previous only the turns this connection has not
        seen. Heartbeats come every 10 s. A move_response must echo request_id, or it is dropped. An
        illegal move forfeits.
      operationId: openEngineSession
      parameters:
        - $ref: "#/components/parameters/GameId"
        - description: The game token from the gameStart line.
          in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "101":
          description: Switching protocols; the engine session is open.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Unknown game, or a game token that is expired or rotated.
      security: []
      summary: Open the per-game engine session.
      tags:
        - Engine session
...
  /api/games:
    get:
      description: Games in progress, newest first, at most 12, without pagination. Guest games are
        listed; finished games never are.
      operationId: listLiveGames
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/LiveGameEntry"
                type: array
          description: The live games.
      security: []
      summary: List live games.
      tags:
        - Games
    post:
      description: The bot must hold its stream open with open=1, have fewer than 4 live games, and accept
        the clock. The caller, a user or guest, may hold 3 live games and create one every 60 s. The
        server draws sides and places the opening; the bot receives gameStart. A game against a
        guest is unrated.
      operationId: createGame
      requestBody:
        content:
          application/json:
            schema:
              properties:
                bot:
                  pattern: ^[A-Za-z][A-Za-z0-9_-]{0,28}[A-Za-z0-9]$
                  type: string
                openingPlies:
                  allOf:
                    - $ref: "#/components/schemas/OpeningPlies"
                    - default: 5
                timeControl:
                  $ref: "#/components/schemas/TimeControl"
              required:
                - bot
                - timeControl
              type: object
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSnapshot"
          description: The new game's snapshot.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameCreateError"
          description: Validation failed (bad_request), the caller is at its live-game cap (human_busy) or
            inside the creation cooldown (game_cooldown), or the bot is not open (not_open),
            excludes the clock (clock_not_accepted), or is at its game cap (bot_busy).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Delisted"
        "503":
          $ref: "#/components/responses/Paused"
      security:
        - sessionCookie: []
      summary: Start a game against a bot.
      tags:
        - Games
  /api/games/{gameId}:
    get:
      description: Board, turn, and clock in one read; a finished game carries its result. Anyone may read
        a game, a guest's while the guest's session lives.
      operationId: getGameSnapshot
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSnapshot"
          description: The game's snapshot.
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - sessionCookie: []
        - {}
      summary: Read any game.
      tags:
        - Games
  /api/games/{gameId}/events:
    get:
      description: "Server-sent events: on open a snapshot, then a turn event per applied turn, then a
        finish event, and the stream closes. A finished game sends its snapshot and closes. A
        comment line comes every 10 s. Events carry no ids and nothing replays: a reconnect gets a
        fresh snapshot."
      operationId: watchGame
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/GameEvent"
          description: The event stream.
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatcherLimitError"
          description: Watchers without a seat are capped at 50 per game and 500 in total; retry after 30 s. A
            seated caller is never refused.
          headers:
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
      security:
        - sessionCookie: []
        - {}
      summary: Watch a game live.
      tags:
        - Games
  /api/games/{gameId}/move:
    post:
      description: Places both stones of the caller's turn. A win on the first placement ends the game,
        and the second is not applied. An illegal move answers 400 and forfeits nothing.
      operationId: playHumanMove
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        content:
          application/json:
            schema:
              properties:
                cells:
                  items:
                    properties:
                      x:
                        type: integer
                      y:
                        type: integer
                    required:
                      - x
                      - y
                    type: object
                  type: array
              required:
                - cells
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSnapshot"
          description: The snapshot after the move.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveError"
          description: Validation failed (bad_request), it is not the caller's turn (not_your_turn), a cell is
            taken (cell_occupied) or out of range (out_of_range), or the game is over (game_over).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - sessionCookie: []
      summary: Play the two placements of the caller's turn.
      tags:
        - Games
  /api/games/{gameId}/resign:
    post:
      description: The opponent wins with reason surrender.
      operationId: resignHumanGame
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSnapshot"
          description: The finished snapshot.
        "400":
          $ref: "#/components/responses/GameOver"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - sessionCookie: []
      summary: Resign a game the caller plays in.
      tags:
        - Games
...